kbuild: detect missing header include guard
Our convention is to surround the whole of header content with include
guard. This avoids the same header being parsed again and again in case
it is included multiple times.

In fact, the include guard is missing in several places, or broken due
to the mismatch between #ifndef and #define.

The recently added header-test-y syntax allows the comprehensive sanity
checks of headers. This commit adds another check; if include guard is
missing, the header will fail to build due to redefinition of something.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>