modpost: write [NO CRC] to Module.symvers when CRC is invalid
When CONFIG_MODVERSION is disabled, there is no CRC bound to exported
symbols.

write_dump() write 0x00000000 to the CRC field in Module.symvers.

When read_dump() reads it, it treats the 0x00000000 as a valid CRC
and sets sym->crc_valid, which seems a weird behavior.

To clarify the difference between a valid CRC value 0x00000000 and
no CRC, write '[NO CRC]' for the latter.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>