x86: check CONFIG options instead of __arch_um__
If the intention is to check i386/x86_64 excluding UML, we can use
CONFIG options instead.

There are only some users of __arch_um__. This commit replaces them,
then removes the __arch_um__ definition.

The original reason for checking __i386__ / __x86_64__ was perhaps
because lib/raid6/algos.c is built not only for the kernel but also
for the user-space test program.

However, lib/raid6/test/Makefile passes -DCONFIG_X86, -DCONFIG_X86_32,
and -DCONFIG_X86_64 for this case. So, I do not see a good reason to
not use CONFIG options here.

What is confusing is, CONFIG_X86_{32,64} is defined not only in
arch/x86/Kconfig, but also in arch/x86/um/Kconfig. To exlucde UML,
we need to check CONFIG_X86 too.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Acked-by: H. Peter Anvin (Intel) <hpa@zytor.com>